+++ src/scenes/Work/index.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import Proptypes from 'prop-types';
 import { connect } from 'react-redux';
 import { loadWork } from '../../store/work/work';
+import { loadHeaderImage } from '../../store/base/base';
 
 import PageHeader from '../../components/Page Header';
 import { Loader } from '../../components/Loader';
@@ -12,13 +13,16 @@ const mapDispatchToProps = dispatch => {
   return {
     loadWork: () => {
       dispatch(loadWork());
+    },
+    loadHeaderImage: () => {
+      dispatch(loadHeaderImage('2iuUpTaObaSCw0kcya8Ci'));
     }
   };
 };
 
 // Maps Redux state to props
 function mapStateToProps(state) {
-  return { work: state.work };
+  return { headerImage: state.base, work: state.work };
 }
 
 class Work extends React.Component {
@@ -29,7 +33,10 @@ class Work extends React.Component {
   }
 
   componentDidMount() {
-    const { loadWork } = this.props;
+    const { loadHeaderImage, loadWork } = this.props;
+
+    // Dispatches headerImage()
+    loadHeaderImage();
 
     // Dispatches loadWork()
     loadWork();
@@ -37,10 +44,12 @@ class Work extends React.Component {
 
   render() {
     const { loading, work } = this.props.work;
-    
-    return <div className=" bg-gray">
+    const { headerImage } = this.props.headerImage;
+
+    return (
+      <div className=" bg-gray">
         <h1>Work</h1>
-        <PageHeader headerImage="https://source.unsplash.com/xtd3zYWxEs4/1600x900">
+        <PageHeader headerImage={headerImage}>
           <h1>Work</h1>
         </PageHeader>
         <div className="work-page container">
@@ -51,8 +60,8 @@ class Work extends React.Component {
                 <p>
                   Lorem ipsum dolor sit, amet consectetur adipisicing elit.
                   Quibusdam odit, soluta voluptatibus minus omnis obcaecati
-                  eveniet officia ipsum molestias animi, vitae dicta atque
-                  in maiores vero temporibus odio. Nesciunt, corporis.
+                  eveniet officia ipsum molestias animi, vitae dicta atque in
+                  maiores vero temporibus odio. Nesciunt, corporis.
                 </p>
               </div>
             </div>
@@ -61,38 +70,44 @@ class Work extends React.Component {
                 <h1>How we work</h1>
                 <p>
                   Lorem ipsum dolor sit amet consectetur adipisicing elit.
-                  Quasi, quas sit, repellat accusamus molestias vitae vel
-                  illo eaque est placeat et fugit repudiandae hic rem a
-                  temporibus non odio voluptates.
+                  Quasi, quas sit, repellat accusamus molestias vitae vel illo
+                  eaque est placeat et fugit repudiandae hic rem a temporibus
+                  non odio voluptates.
                 </p>
               </div>
             </div>
           </div>
           <div className="work-section row">
-            {loading ? <Loader /> : work.map(({ fields, sys }, i) => (
+            {loading ? (
+              <Loader />
+            ) : (
+              work.map(({ fields, sys }, i) => (
                 <div className="col-12 col-md-4" key={i}>
                   <WorkItem {...fields} {...sys} />
                 </div>
-              ))}
+              ))
+            )}
           </div>
           <div className="work-section row justify-content-center">
             <div className="col-6 text-center">
               <h3>Work With Us</h3>
               <p>
                 Lorem, ipsum dolor sit amet consectetur adipisicing elit. A,
-                maxime, animi consequatur doloremque totam nostrum facilis
-                quo voluptatum ipsam illum quod quaerat neque! Ad dicta unde
-                dolores voluptates veritatis laborum.
+                maxime, animi consequatur doloremque totam nostrum facilis quo
+                voluptatum ipsam illum quod quaerat neque! Ad dicta unde dolores
+                voluptates veritatis laborum.
               </p>
             </div>
           </div>
         </div>
-      </div>;
+      </div>
+    );
   }
 }
 
 Work.propTypes = {
-  loadWork: Proptypes.func
+  loadWork: Proptypes.func,
+  loadHeaderImage: Proptypes.func
 };
 
 export default connect(