+++ src/store/base/types.js
@@ -1,0 +1,4 @@
+// Base types
+
+// Successfully loaded header image
+export const LOAD_HEADER_IMAGE_SUCCESS = 'LOAD_HEADER_IMAGE_SUCCESS';