+++ src/store/initialState.js
@@ -1,5 +1,8 @@
 // Initial Redux state
 export default {
+    base: {
+        headerImage: ''
+    },
     blog: {
         loading: false,
         posts: [],